# test/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/custom-query/rss/{id}": {
            "get": {
                "operationId": "CustomQuery_customQuery",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "key",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "custom-query"
                ]
            }
        },
        "/api/custom-query": {
            "put": {
                "operationId": "CustomQuery_update",
                "summary": "更新 CustomQuery",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCustomQuery"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "响应成功"
                    },
                    "201": {
                        "description": "创建成功"
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "custom-query"
                ]
            },
            "get": {
                "operationId": "CustomQuery_find",
                "summary": "查找所有记录",
                "parameters": [
                    {
                        "name": "query",
                        "required": false,
                        "in": "query",
                        "description": "Query options",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FindCustomQuery"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "custom-query"
                ]
            },
            "post": {
                "operationId": "CustomQuery_create",
                "summary": "创建记录",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCustomQuery"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomQuery"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "custom-query"
                ]
            }
        },
...
        "/api/custom-query/{id}": {
            "get": {
                "operationId": "CustomQuery_findOne",
                "summary": "查找记录",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomQuery"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "custom-query"
                ]
            },
            "delete": {
                "operationId": "CustomQuery_delete",
                "summary": "删除记录",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomQuery"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "custom-query"
                ]
            }
        },
...
    },
    "info": {
        "title": "RSS Impact server docs",
        "description": "RSS Impact server docs",
        "version": "0.1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "connect.sid"
            }
        },
        "schemas": {
...
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "title": "ID",
                        "example": 1
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string",
                        "title": "创建时间",
                        "example": "2023-12-31T16:00:00.000Z"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string",
                        "title": "更新时间",
                        "example": "2023-12-31T16:00:00.000Z"
                    },
                    "username": {
                        "type": "string",
                        "title": "用户名",
                        "example": "admin"
                    },
                    "password": {
                        "type": "string",
                        "title": "密码",
                        "example": "123456"
                    },
                    "email": {
                        "type": "string",
                        "title": "邮箱",
                        "example": "admin@example.com"
                    },
                    "emailVerified": {
                        "type": "boolean",
                        "title": "邮箱已验证",
                        "example": true
                    },
                    "disablePasswordLogin": {
                        "type": "boolean",
                        "title": "禁用密码登录",
                        "example": false
                    },
                    "avatar": {
                        "type": "string",
                        "title": "头像",
                        "example": "URL_ADDRESS"
                    },
                    "roles": {
                        "title": "角色",
                        "example": [
                            "admin"
                        ],
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "accessToken": {
                        "type": "string",
                        "title": "接口访问令牌",
                        "description": "接口访问令牌，部分情况可替代账号密码",
                        "example": "rss-impact:fef26d6e-040f-4a7b-8d6a-4e4f12e107b6"
                    },
                    "auth0Id": {
                        "type": "string",
                        "title": "Auth0 ID",
                        "description": "绑定的 auth0 账号",
                        "example": "github|114514"
                    }
                },
                "required": [
                    "id",
                    "createdAt",
                    "updatedAt",
                    "username",
                    "email",
                    "emailVerified",
                    "disablePasswordLogin",
                    "roles",
                    "accessToken"
                ]
            },
...
            "FilterOut": {
                "type": "object",
                "properties": {
                    "link": {
                        "type": "string",
                        "minLength": 0,
                        "maxLength": 2048,
                        "title": "排除链接",
                        "example": "url1|url2"
                    },
                    "title": {
                        "type": "string",
                        "minLength": 0,
                        "maxLength": 256,
                        "title": "排除标题",
                        "example": "标题1|标题2"
                    },
                    "content": {
                        "type": "string",
                        "minLength": 0,
                        "maxLength": 2048,
                        "title": "排除正文",
                        "example": "正文1|正文2"
                    },
                    "summary": {
                        "type": "string",
                        "minLength": 0,
                        "maxLength": 1024,
                        "title": "排除总结",
                        "example": "总结1|总结2"
                    },
                    "author": {
                        "type": "string",
                        "minLength": 0,
                        "maxLength": 256,
                        "title": "排除作者",
                        "example": "author1|author2"
                    },
                    "categories": {
                        "type": "string",
                        "minLength": 0,
                        "maxLength": 256,
                        "title": "排除分类",
                        "description": "分类正则中有一个对得上就排除",
                        "example": "tag1|tag2"
                    },
                    "enclosureUrl": {
                        "type": "string",
                        "minLength": 0,
                        "maxLength": 2048,
                        "title": "排除附件URL",
                        "example": "url1|url2"
                    },
                    "enclosureType": {
                        "type": "string",
                        "minLength": 0,
                        "maxLength": 256,
                        "title": "排除附件类型",
                        "example": "type1|type2"
                    }
                }
            },
...
            "Category": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "title": "ID",
                        "example": 1
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string",
                        "title": "创建时间",
                        "example": "2023-12-31T16:00:00.000Z"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string",
                        "title": "更新时间",
                        "example": "2023-12-31T16:00:00.000Z"
                    },
                    "userId": {
                        "type": "number",
                        "title": "所属用户",
                        "example": 1
                    },
                    "user": {
                        "title": "所属用户",
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ]
                    },
                    "name": {
                        "type": "string",
                        "title": "名称",
                        "example": "分类A"
                    },
                    "description": {
                        "type": "string",
                        "title": "简介",
                        "example": "分类A"
                    },
                    "feeds": {
                        "title": "订阅链接",
                        "example": [],
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Feed"
                        }
                    }
                },
                "required": [
                    "id",
                    "createdAt",
                    "updatedAt",
                    "userId",
                    "user",
                    "name",
                    "feeds"
                ]
            },
...
            "CustomQuery": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "title": "ID",
                        "example": 1
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string",
                        "title": "创建时间",
                        "example": "2023-12-31T16:00:00.000Z"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string",
                        "title": "更新时间",
                        "example": "2023-12-31T16:00:00.000Z"
                    },
                    "userId": {
                        "type": "number",
                        "title": "所属用户",
                        "example": 1
                    },
                    "user": {
                        "title": "所属用户",
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ]
                    },
                    "name": {
                        "type": "string",
                        "title": "名称",
                        "example": "查询A"
                    },
                    "scope": {
                        "type": "object",
                        "enum": [
                            "all",
                            "category",
                            "feed"
                        ],
                        "title": "查询范围",
                        "description": "指定分类和指定订阅的配置互斥，只按照本项指定的范围查询",
                        "example": "all"
                    },
                    "categories": {
                        "title": "指定分类",
                        "description": "支持选择多个分类",
                        "example": [],
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "feeds": {
                        "title": "指定订阅",
                        "description": "支持选择多个订阅",
                        "example": [],
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Feed"
                        }
                    },
                    "format": {
                        "type": "object",
                        "enum": [
                            "rss2.0",
                            "atom",
                            "json"
                        ],
                        "title": "输出格式",
                        "example": "rss2.0"
                    },
                    "useAiSummary": {
                        "type": "boolean",
                        "title": "使用 AI 总结",
                        "description": "如果是，则用 AI 总结替换原本的总结"
                    },
                    "appendAiSummary": {
                        "type": "boolean",
                        "title": "增加 AI 总结",
                        "description": "如果是，则将 AI 总结 增加 到正文前，以方便通过 RSS 阅读器阅读"
                    },
                    "key": {
                        "type": "string",
                        "title": "访问秘钥",
                        "description": "通过访问秘钥即可无需登录访问 RSS 订阅。一旦泄露，请立即修改！",
                        "example": "custom-query-key:2c28d0b6-47db-43a4-aff4-439edbe29200"
                    },
                    "url": {
                        "type": "string",
                        "title": "输出路径"
                    },
                    "filter": {
                        "title": "过滤条件",
                        "description": "保留想要的内容，必须符合全部条件才保留。支持通过正则表达式过滤。留空的规则不会过滤。",
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Filter"
                            }
                        ]
                    },
                    "filterout": {
                        "title": "排除条件",
                        "description": "去掉不要的内容，有一个条件符合就排除。支持通过正则表达式排除。留空的规则不会排除。",
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/FilterOut"
                            }
                        ]
                    }
                },
                "required": [
                    "id",
                    "createdAt",
                    "updatedAt",
                    "userId",
                    "user",
                    "name",
                    "scope",
                    "format",
                    "useAiSummary",
                    "appendAiSummary",
                    "key",
                    "filter",
                    "filterout"
                ]
            },
...
            "UpdateCustomQuery": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "title": "ID",
                        "example": 1
                    },
                    "userId": {
                        "type": "number",
                        "title": "所属用户",
                        "example": 1
                    },
                    "user": {
                        "title": "所属用户",
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ]
                    },
                    "name": {
                        "type": "string",
                        "title": "名称",
                        "example": "查询A"
                    },
                    "scope": {
                        "type": "object",
                        "enum": [
                            "all",
                            "category",
                            "feed"
                        ],
                        "title": "查询范围",
                        "description": "指定分类和指定订阅的配置互斥，只按照本项指定的范围查询",
                        "example": "all"
                    },
                    "categories": {
                        "title": "指定分类",
                        "description": "支持选择多个分类",
                        "example": [],
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "feeds": {
                        "title": "指定订阅",
                        "description": "支持选择多个订阅",
                        "example": [],
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Feed"
                        }
                    },
                    "format": {
                        "type": "object",
                        "enum": [
                            "rss2.0",
                            "atom",
                            "json"
                        ],
                        "title": "输出格式",
                        "example": "rss2.0"
                    },
                    "useAiSummary": {
                        "type": "boolean",
                        "title": "使用 AI 总结",
                        "description": "如果是，则用 AI 总结替换原本的总结"
                    },
                    "appendAiSummary": {
                        "type": "boolean",
                        "title": "增加 AI 总结",
                        "description": "如果是，则将 AI 总结 增加 到正文前，以方便通过 RSS 阅读器阅读"
                    },
                    "url": {
                        "type": "string",
                        "title": "输出路径"
                    },
                    "filter": {
                        "title": "过滤条件",
                        "description": "保留想要的内容，必须符合全部条件才保留。支持通过正则表达式过滤。留空的规则不会过滤。",
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Filter"
                            }
                        ]
                    },
                    "filterout": {
                        "title": "排除条件",
                        "description": "去掉不要的内容，有一个条件符合就排除。支持通过正则表达式排除。留空的规则不会排除。",
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/FilterOut"
                            }
                        ]
                    }
                }
            },
            "FindCustomQuery": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CustomQuery"
                        }
                    },
                    "total": {
                        "type": "number"
                    },
                    "lastPage": {
                        "type": "number"
                    },
                    "currentPage": {
                        "type": "number"
                    }
                },
                "required": [
                    "data",
                    "total",
                    "lastPage",
                    "currentPage"
                ]
            },
            "CreateCustomQuery": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "number",
                        "title": "所属用户",
                        "example": 1
                    },
                    "user": {
                        "title": "所属用户",
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ]
                    },
                    "name": {
                        "type": "string",
                        "title": "名称",
                        "example": "查询A"
                    },
                    "scope": {
                        "type": "object",
                        "enum": [
                            "all",
                            "category",
                            "feed"
                        ],
                        "title": "查询范围",
                        "description": "指定分类和指定订阅的配置互斥，只按照本项指定的范围查询",
                        "example": "all"
                    },
                    "categories": {
                        "title": "指定分类",
                        "description": "支持选择多个分类",
                        "example": [],
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "feeds": {
                        "title": "指定订阅",
                        "description": "支持选择多个订阅",
                        "example": [],
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Feed"
                        }
                    },
                    "format": {
                        "type": "object",
                        "enum": [
                            "rss2.0",
                            "atom",
                            "json"
                        ],
                        "title": "输出格式",
                        "example": "rss2.0"
                    },
                    "useAiSummary": {
                        "type": "boolean",
                        "title": "使用 AI 总结",
                        "description": "如果是，则用 AI 总结替换原本的总结"
                    },
                    "appendAiSummary": {
                        "type": "boolean",
                        "title": "增加 AI 总结",
                        "description": "如果是，则将 AI 总结 增加 到正文前，以方便通过 RSS 阅读器阅读"
                    },
                    "url": {
                        "type": "string",
                        "title": "输出路径"
                    },
                    "filter": {
                        "title": "过滤条件",
                        "description": "保留想要的内容，必须符合全部条件才保留。支持通过正则表达式过滤。留空的规则不会过滤。",
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Filter"
                            }
                        ]
                    },
                    "filterout": {
                        "title": "排除条件",
                        "description": "去掉不要的内容，有一个条件符合就排除。支持通过正则表达式排除。留空的规则不会排除。",
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/FilterOut"
                            }
                        ]
                    }
                },
                "required": [
                    "userId",
                    "user",
                    "name",
                    "scope",
                    "format",
                    "useAiSummary",
                    "appendAiSummary",
                    "filter",
                    "filterout"
                ]
            },
...
        }
    }
}